Module Tezt.Background

val register : unit Lwt.t -> unit

Register a promise that will run in the background.

After a test runs, Test.run waits for all registered promises to finish.

If a registered promise raises an exception which is not Lwt.Canceled, the current test fails immediately, but Test.run still waits for other background promises to finish.

Make sure that the promise you register eventually resolves. If it doesn't, stop (and thus Test.run, which calls stop) will hang forever.

Calls to register when no test is running result in an error.

val start : (exn -> unit) -> unit

Allow calls to register until stop is called.

If a promise that is later registered is rejected by an exception which is not Lwt.Canceled, start calls its argument. This may occur several times, including during stop.

raises [Invalid_arg]

if calls to register are already allowed, i.e. if start has already been called without a corresponding stop.

Don't call this directly, it is called by Test.run.

val stop : unit -> unit Lwt.t

Let all registered promises resolve, then stop allowing calls to register.

The promise returned by stop is never rejected except if you cancel it. If you do cancel it, you may have to call stop again though as it is not guaranteed that all background promises have also been canceled. In other words, canceling the promise returned by stop is probably a bad idea.

Calling stop when calls to register are not allowed has no effect, i.e. you can call stop even with no corresponding start.

Don't call this directly, it is called by Test.run.