Module Tezt.Cli

Command-line interface.

type log_level =

| Quiet

| Error

| Warn

| Report

| Info

| Debug

Log levels for standard output.

The list below is sorted from the most quiet level to the most verbose level.

Absolutely no log has log level Quiet. In other words, setting log level Quiet inhibits all logs.

Error logs are about errors which imply that the current test failed. This includes messages given to Test.fail and uncaught exceptions.

Warn logs are about errors that do not cause the current test to fail. This includes failure to clean up temporary files, for instance.

Report logs are informational messages that report the result of the current test. They tell the user whether the test was successful or not. They may also include information about how to re-run the test.

Info logs are informational messages that summarize what the test is doing. They tell the user that a particular milestone was reached. In tests, it is a good idea to log Info messages at significant checkpoints.

Debug logs give more details about exactly what is happening. They include external process outputs, exit codes, and signals which are sent.

Additionally, some flags such as --commands and --list cause some information to be printed unconditionally, even with --quiet. Such kind of output is not considered to be log messages.

type temporary_file_mode =

| Delete

| Delete_if_successful

| Keep

What to do with temporary files after the test is finished.

type loop_mode =

| Infinite

| Count of int

How many times to loop.

type options = {

color : bool;

log_level : log_level;

log_file : string option;

log_buffer_size : int;

commands : bool;

temporary_file_mode : temporary_file_mode;

keep_going : bool;

files_to_run : string list;

tests_to_run : string list;

tests_not_to_run : string list;

tags_to_run : string list;

tags_not_to_run : string list;

list : [ `Ascii_art | `Tsv ] option;

global_timeout : float option;

test_timeout : float option;

reset_regressions : bool;

loop_mode : loop_mode;

time : bool;

starting_port : int;

record : string option;

from_records : string list;

job : (int * int) option;

job_count : int;

suggest_jobs : bool;

junit : string option;

}

Command-line options.

val options : options

Values for command-line options.