Module Tezt__Progress

Progress indicator for Tezt

type t

A structure to track the progress of a test suite.

val create : total:int -> t

create ~total creates a progress indicator for an expected total numbers of tests.

val update : has_failed:bool -> t -> unit

update ~has_failed progress updates progress state with a test, usually the last one, which has either failed ~has_failed:true or not ~has_failed:false.

val pp : Stdlib.Format.formatter -> t -> unit

Pretty-printer for progress indicator.