Module Tezt__Regression

Run Tezt regression tests and capture their output.

NOTE: consider using dune cram when it becomes available: https://dune.readthedocs.io/en/stable/tests.html#cram-tests

val register : __FILE__:string -> title:string -> tags:string list -> output_file:string -> ?⁠regression_output_path:string -> (unit -> unit Lwt.t) -> unit

Register a regression test.

This function is a wrapper around Test.register. It adds the "regression" tag to the tags list provided by the argument.

The output_file is a filename, where the output of the test is stored and read to be compared on subsequent runs. The output file with ".out" filename extension added is placed in the regression_output_path, which is by default "tezt/_regressions".

val capture : string -> unit

Capture some output of a regression test.

Call this to record a string into the output_file given to register. A newline character \n will be added after it.

This function only records its argument when called while a regression test is running, i.e. from the body of Regression.register. If you call it outside of Regression.register, it has no effect. So you can define a function that captures and use it with or without regression testing.

A typical use is to define custom process hooks that substitute non-deterministic parts of the output with deterministic ones. See also hooks.

val hooks : Tezt.Process.hooks

Hooks that enable regression testing when attached to a process ran from a registered regression test function.

The hooks will capture the spawned command, its arguments and the output of its execution into the registered output_file.